umaten/requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
lxml>=4.9.0

# Web UI
Flask>=3.0.0
Flask-SocketIO>=5.3.0
psutil>=5.9.0

# API Clients
anthropic>=0.25.0
python-wordpress-xmlrpc>=2.3

# Data Processing
pyyaml>=6.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Utilities
fake-useragent>=1.4.0
retry>=0.9.2
colorama>=0.4.6
tqdm>=4.66.0

# Logging
loguru>=0.7.0

# Rate Limiting & Delays
ratelimit>=2.2.1

# Proxy Support (optional)
requests[socks]>=2.31.0
